+++ src/main/main.ts
@@ -60,6 +60,7 @@ export interface Config {
   domain: string
   language?: string
   locale?: string
+  webAuthnOrigin?: string
 }
 
 export type ApiClientConfig = {
@@ -135,7 +136,7 @@ export function createClient(creationConfig: Config): Client {
   checkParam(creationConfig, 'domain')
   checkParam(creationConfig, 'clientId')
 
-  const { domain, clientId, language, locale } = creationConfig
+  const { domain, clientId, language, locale, webAuthnOrigin } = creationConfig
 
   const eventManager = createEventManager()
 
@@ -203,7 +204,7 @@ export function createClient(creationConfig: Config): Client {
   )
 
   function addNewWebAuthnDevice(accessToken: string, friendlyName?: string) {
-    return apiClients.then(clients => clients.webAuthn.addNewWebAuthnDevice(accessToken, friendlyName))
+    return apiClients.then(clients => clients.webAuthn.addNewWebAuthnDevice(accessToken, friendlyName, webAuthnOrigin))
   }
 
   function checkSession(options: AuthOptions = {}) {
@@ -275,7 +276,10 @@ export function createClient(creationConfig: Config): Client {
   }
 
   function loginWithWebAuthn(params: LoginWithWebAuthnParams) {
-    return apiClients.then(clients => clients.webAuthn.loginWithWebAuthn(params))
+    return apiClients.then(clients => clients.webAuthn.loginWithWebAuthn({
+      ...params,
+      webAuthnOrigin
+    }))
   }
 
   function logout(params: LogoutParams = {}, revocationParams?: RevocationParams) {
@@ -324,7 +328,10 @@ export function createClient(creationConfig: Config): Client {
   }
 
   function resetPasskeys(params: ResetPasskeysParams) {
-    return apiClients.then(clients => clients.webAuthn.resetPasskeys(params))
+    return apiClients.then(clients => clients.webAuthn.resetPasskeys({
+      ...params,
+      webAuthnOrigin
+    }))
   }
 
   function sendEmailVerification(params: EmailVerificationParams) {
@@ -340,7 +347,10 @@ export function createClient(creationConfig: Config): Client {
   }
 
   function signupWithWebAuthn(params: SignupWithWebAuthnParams, auth?: AuthOptions) {
-    return apiClients.then(clients => clients.webAuthn.signupWithWebAuthn(params, auth))
+    return apiClients.then(clients => clients.webAuthn.signupWithWebAuthn({
+      ...params,
+      webAuthnOrigin
+    }, auth))
   }
 
   function startMfaEmailRegistration(params: StartMfaEmailRegistrationParams) {