+++ src/main/webAuthnClient.ts
@@ -9,13 +9,14 @@ import {
   EmailLoginWithWebAuthnParams,
   encodePublicKeyCredentialCreationOptions,
   encodePublicKeyCredentialRequestOptions,
+  InternalLoginWithWebAuthnParams,
+  InternalSignupWithWebAuthnParams,
   LoginWithWebAuthnParams,
   PhoneNumberLoginWithWebAuthnParams,
   publicKeyCredentialType,
   RegistrationOptions,
   serializeAuthenticationPublicKeyCredential,
   serializeRegistrationPublicKeyCredential,
-  SignupWithWebAuthnParams
 } from './webAuthnService'
 import { ApiClientConfig } from './main'
 import OAuthClient from './oAuthClient'
@@ -36,6 +37,8 @@ type SmsResetPasskeysParams = {
 }
 
 export type ResetPasskeysParams = EmailResetPasskeysParams | SmsResetPasskeysParams
+
+export type InternalResetPasskeysParams = { webAuthnOrigin?: string } & ResetPasskeysParams
 
 /**
  * Identity Rest API Client
@@ -78,10 +81,10 @@ export default class WebAuthnClient {
     return (credentials as PublicKeyCredential).type === publicKeyCredentialType
   }
 
-  addNewWebAuthnDevice(accessToken: string, friendlyName?: string): Promise<void> {
+  addNewWebAuthnDevice(accessToken: string, friendlyName?: string, webAuthnOrigin?: string): Promise<void> {
     if (window.PublicKeyCredential) {
       const body = {
-        origin: window.location.origin,
+        origin: webAuthnOrigin || window.location.origin,
         friendlyName: friendlyName || window.navigator.platform
       }
 
@@ -111,11 +114,11 @@ export default class WebAuthnClient {
     }
   }
 
-  resetPasskeys(params: ResetPasskeysParams): Promise<void> {
+  resetPasskeys(params: InternalResetPasskeysParams): Promise<void> {
     if (window.PublicKeyCredential) {
       const body = {
         ...params,
-        origin: window.location.origin,
+        origin: params.webAuthnOrigin || window.location.origin,
         friendlyName: params.friendlyName || window.navigator.platform
       }
 
@@ -155,16 +158,16 @@ export default class WebAuthnClient {
     return typeof (params as DiscoverableLoginWithWebAuthnParams).conditionalMediation !== 'undefined'
   }
 
-  private buildWebAuthnParams(params: LoginWithWebAuthnParams): Promise<LoginWithWebAuthnQueryParams> {
+  private buildWebAuthnParams(params: InternalLoginWithWebAuthnParams): Promise<LoginWithWebAuthnQueryParams> {
     const body = this.isDiscoverable(params)
       ? {
           clientId: this.config.clientId,
-          origin: window.location.origin,
+          origin: params.webAuthnOrigin || window.location.origin,
           scope: resolveScope(params.auth, this.config.scope)
         }
       : {
           clientId: this.config.clientId,
-          origin: window.location.origin,
+          origin: params.webAuthnOrigin || window.location.origin,
           scope: resolveScope(params.auth, this.config.scope),
           email: (params as EmailLoginWithWebAuthnParams).email,
           phoneNumber: (params as PhoneNumberLoginWithWebAuthnParams).phoneNumber
@@ -181,7 +184,7 @@ export default class WebAuthnClient {
     })
   }
 
-  loginWithWebAuthn(params: LoginWithWebAuthnParams): Promise<AuthResult> {
+  loginWithWebAuthn(params: InternalLoginWithWebAuthnParams): Promise<AuthResult> {
     if (!window.PublicKeyCredential) {
       return Promise.reject(new Error('Unsupported WebAuthn API'))
     }
@@ -231,10 +234,10 @@ export default class WebAuthnClient {
     return this.http.remove<void>(`${this.registrationUrl}/${deviceId}`, { accessToken })
   }
 
-  signupWithWebAuthn(params: SignupWithWebAuthnParams, auth?: AuthOptions): Promise<AuthResult> {
+  signupWithWebAuthn(params: InternalSignupWithWebAuthnParams, auth?: AuthOptions): Promise<AuthResult> {
     if (window.PublicKeyCredential) {
       const body = {
-        origin: window.location.origin,
+        origin: params.webAuthnOrigin || window.location.origin,
         clientId: this.config.clientId,
         friendlyName: params.friendlyName || window.navigator.platform,
         profile: params.profile,