+++ CHANGELOG.md
@@ -7,6 +7,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 - Add support for account recovery and passkey reset
+- Make WebAuthn origin configurable
 
 ## [1.32.2] - 2023-03-21
 - Add support for discoverable passkey login